User defined colours for pie controls (1/2)
Although CyanogenMod's colour scheme is normally fixed and only
changed by themes, this commit allows the user to change some of
the colours of the pie.

The rationale behind this is as follows:
Normally UI controls both - foreground and background colour of
each UI element. In case of the pie controls this is not possible,
pie controls are possibly overlayed in all screens of the UI and
all custom installed applications. This makes it hard to come up
with a universal colour scheme for pie controls.
In other words the user perceived "average" background
colour *) varies from user to user and thus universal colours
for pie controls are hard to pick.

*) It showed, that the background of the home
screen is major factor to this. Due to the fact that the
background is user settable, the colour range here is almost
infinte.

This commit adds:
* A user can change the following colours from within the Settings
  application:
  Normal button background, Selected button background,
  Long pressed button background, Icon colour
* Each colour can be reseted to the system default colour.

There is a seperate commit for the Settings application.

Change-Id: I34975b02999303d4ee479918d6e22a95e940a6ef