fb: fix rounding issues and take density into account
we are a multiDpi rom...and on 4.2 and 4.3 already the text
in the circle battery icon was on lower dpi a small bit up. The original
author compensated this with just a +1 to the Y value. Thats not the right
way...calculate the current dpi into the offset and all is well.

We want a perfect 4.4...so this nitpicks even if we are talking about
0.5dp maybe should be fixed

Change-Id: I835aa20b82aea980eb79a4772adccf30b2d6f79f